python/requirements.txt
requests>=2.31.0
lxml>=5.1.0
beautifulsoup4>=4.12.0
SPARQLWrapper>=2.0.0
rdflib>=7.0.0
weaviate-client>=4.9.0
voyageai>=0.3.0
google-generativeai>=0.8.0
